requirements.txt
setuptools
Flask==0.12.2
configparser
tornado
sqlalchemy==1.1.15
schedule
requests
openpyxl
image
psycopg2
PyMySQL
python-dateutil
pyfcm
pytz
xhtml2pdf
PyJWT
cx_oracle
fpdf
xlwt
numpy
matplotlib
dropbox
psutil
